engine_rocks: Move db.rs to engine.rs for consistency
