+++ components/engine_rocks/src/cf_handle.rs
@@ -1,7 +1,7 @@
 // Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
 
 use crate::cf_options::RocksCFOptions;
-use crate::db::Rocks;
+use crate::engine::Rocks;
 use engine_traits::CFHandle;
 use engine_traits::CFHandleExt;
 use engine_traits::Result;
